Do not backup if not exists

+++ scripts/rbenv.sh
@@ -3,7 +3,9 @@
 set -u
 
 if command -v git &>/dev/null; then
-  mv "$HOME/.rbenv" "$HOME/.rbenv.bak"
+  if [ -e "$HOME/.rbenv" ]; then
+    mv "$HOME/.rbenv" "$HOME/.rbenv.bak"
+  fi
 
   git clone "https://github.com/rbenv/rbenv.git" "$HOME/.rbenv"
   git clone "https://github.com/rbenv/ruby-build.git" "$HOME/.rbenv/plugins/ruby-build"
